requirements.txt
Pillow==10.0.1
pycryptodome==3.19.0
xxhash==3.4.1
